Add category management and UI enhancements
- Prevent deletion of categories assigned to tasks by adding a validation check in `DeleteCategoryCommandHandler`.
- Introduce `CheckIfWasAssignedToATaskAsync` in `ICategoriesRepository` and implement it in `CategoriesRepository`.
- Enable category selection for tasks in `CreateTaskPage` and `EditTaskPage` using a new `CategorySelector` component.
- Add `CategoryId` property to `BaseTaskCommand` and update related mappings in `TaskMappingProfile`.
- Enhance `Table.razor` to display categories with a `SmartChip` component and refactor redundant logic.
- Improve error handling in `DeleteDialog.razor` and `Table.razor` with user feedback via snackbars.
- Introduce `SmartChip` for dynamic category display and ensure text readability with contrast calculations.
- Refactor and modularize components for better maintainability and extensibility.

+++ ToDoApp.Application/Features/Categories/Commands/Delete/DeleteCategoryCommandHandler.cs
@@ -10,6 +10,15 @@ public class DeleteCategoryCommandHandler(ICategoriesRepository repository) : IR
     {
         try
         {
+            bool wasAssignedToATask = await repository.CheckIfWasAssignedToATaskAsync(command.Id);
+
+            if (wasAssignedToATask)
+            {
+                throw new Exception(@"Category was assigned to a task. 
+                                    Please remove the category from the task before deleting it.
+                                    You can do this by using the Edit Task command.");
+            }
+
             await repository.DeleteByIdAsync(command.Id);
         }
         catch (Exception)
